requirements.txt
asgiref==3.7
Django==5.0
django-crispy-forms==1.12.0
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.2
